+++ include/SoccerPlayer.hpp
@@ -38,8 +38,8 @@ class SoccerPlayer : public SoccerPlayerAttributes {
     	int getStrength() const override;
     	void setStrength(int value) override;
 
-    	int getResistance() const override;
-    	void setResistance(int value) override;
+    	float getResistance() const override;
+    	void setResistance(float value) override;
 	
 		void informations() const override;
  };
